update profiles to ready and complete imports in same task
