+++ core/src/modules/ops/profile.ts
@@ -12,6 +12,7 @@ import { waitForLock } from "../locks";
 import { ProfilePropertyOps } from "./profileProperty";
 import { CLS } from "../../modules/cls";
 import { GroupRule } from "../../models/GroupRule";
+import { Import } from "../../models/Import";
 
 export interface ProfilePropertyType {
   [key: string]: {
@@ -690,9 +691,12 @@ export namespace ProfileOps {
 
   /**
    * Find profiles that are not ready but whose properties are and make them ready.
-   * Task `profile:completeImport` will then process the profiles.
+   * Then, process the related imports.
    */
-  export async function makeReady(limit = 100, toExport = true) {
+  export async function makeReadyAndCompleteImports(
+    limit = 100,
+    toExport = true
+  ) {
     let profiles: Profile[] = await api.sequelize.query(
       `
     SELECT "id" from "profiles" where "state" = 'pending'
@@ -723,11 +727,84 @@ export namespace ProfileOps {
 
     if (profiles.length === 0) return [];
 
-    await CLS.enqueueTask("profile:completeImport", {
-      profileIds: profiles.map((p) => p.id),
-      toExport,
-    });
+    await completeProfileImports(
+      profiles.map((p) => p.id),
+      toExport
+    );
 
     return profiles;
+  }
+
+  async function completeProfileImports(
+    profileIds: string[],
+    toExport: boolean
+  ) {
+    const properties = await Property.findAllWithCache();
+    const profiles = await Profile.findAll({
+      where: {
+        id: { [Op.in]: profileIds },
+      },
+      include: [
+        { model: ProfileProperty, required: true },
+        { model: Import, required: false, where: { profileUpdatedAt: null } },
+      ],
+    });
+    if (profiles.length === 0) return;
+
+    for (const profile of profiles) {
+      const mergedValues = {};
+      const imports = profile.imports;
+
+      for (const _import of imports.sort(
+        (a, b) => a.createdAt.getTime() - b.createdAt.getTime()
+      )) {
+        const data = _import.data;
+        for (const key in data) {
+          // only if we still have property
+          if (properties.find((r) => r.key === key)) {
+            mergedValues[key] = data[key];
+          }
+        }
+      }
+
+      if (Object.keys(mergedValues).length > 0) {
+        await profile.addOrUpdateProperties(mergedValues);
+        delete profile.profileProperties; // will be reloaded
+      }
+    }
+
+    const memberships = await ProfileOps.updateGroupMemberships(profiles);
+    const now = new Date();
+
+    for (const profile of profiles) {
+      const imports = profile.imports;
+      if (imports.length > 0) {
+        const newProfileProperties = await profile.simplifiedProperties();
+        const newGroupIds = Object.keys(memberships[profile.id]).filter(
+          (groupId) => memberships[profile.id][groupId] === true
+        );
+
+        await Import.update(
+          {
+            newProfileProperties: newProfileProperties,
+            profileUpdatedAt: now,
+            newGroupIds: newGroupIds,
+            groupsUpdatedAt: now,
+            exportedAt: toExport ? undefined : now, // we want to indicate that the import's lifecycle is complete
+          },
+          {
+            where: { id: { [Op.in]: imports.map((i) => i.id) } },
+          }
+        );
+      }
+
+      // TODO call this from a recurring task
+      if (toExport) {
+        await CLS.enqueueTask("profile:export", {
+          profileId: profile.id,
+          force: false,
+        });
+      }
+    }
   }
 }
